line chart not showing fixed
