+++ app/src/main/java/com/iranmobiledev/moodino/ui/states/StatsFragment.kt
@@ -10,12 +10,16 @@ import android.view.ViewGroup
 import android.widget.FrameLayout
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.core.content.ContextCompat
 import com.github.mikephil.charting.components.XAxis
+import com.github.mikephil.charting.data.Entry
+import com.github.mikephil.charting.data.LineData
 import com.github.mikephil.charting.data.LineDataSet
+import com.github.mikephil.charting.formatter.ValueFormatter
 import com.iranmobiledev.moodino.R
 import com.iranmobiledev.moodino.base.BaseFragment
-import com.iranmobiledev.moodino.databinding.DaysInARowCardBinding
 import com.iranmobiledev.moodino.databinding.FragmentStatsBinding
+import com.iranmobiledev.moodino.ui.calendar.calendarpager.formatNumber
 import com.iranmobiledev.moodino.ui.states.viewmodel.StatsFragmentViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
@@ -64,7 +68,6 @@ class StatsFragment : BaseFragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
         initDayInRowCard()
         initLineChartCard()
     }
@@ -82,6 +85,19 @@ class StatsFragment : BaseFragment() {
 
         setupWeekDays()
         setupDaysStatus()
+    }
+
+
+    private fun initLineChartCard() {
+        model.initLineChart()
+        model.lineChartEntries.observe(viewLifecycleOwner){
+            val dataSet = setupLineChart(it)
+            customizeLineChart(dataSet)
+            binding.moodsLineChart.apply {
+                notifyDataSetChanged()
+                invalidate()
+            }
+        }
     }
 
     private fun setupWeekDays() {
@@ -113,18 +129,27 @@ class StatsFragment : BaseFragment() {
         }
     }
 
-    private fun initLineChartCard() {
-        model.lineChartEntries.observe(viewLifecycleOwner) {
-            model.initLineChart(it, requireContext())
-        }
-
-        customizeLineChart()
+    private fun setupLineChart(entries: List<Entry>): LineDataSet {
+        var dataSet = LineDataSet(entries, "moods")
+        var data = LineData(dataSet)
+        binding.moodsLineChart.data = data
+        return dataSet
     }
 
-    private fun customizeLineChart() {
+    private fun customizeLineChart(dataSet: LineDataSet) {
+        dataSet.apply {
+            color = Color.RED
+            lineWidth = 2f
+            highLightColor = R.color.pink
+            setDrawFilled(true)
+            fillDrawable = ContextCompat.getDrawable(requireContext(), R.drawable.chart_gradient)
+            circleHoleColor = (Color.WHITE)
+            setCircleColor(Color.RED);
+            valueTextColor = Color.WHITE
+            valueTextSize = 1f
+        }
+
         binding.moodsLineChart.apply {
-            data = lineData
-            invalidate()
             description.isEnabled = false
             legend.isEnabled = false
             axisRight.isEnabled = false
