+++ app/src/main/java/com/iranmobiledev/moodino/utlis/ChartValueFormatter.kt
@@ -1,0 +1,11 @@
+package com.iranmobiledev.moodino.utlis
+
+import com.github.mikephil.charting.components.AxisBase
+import com.github.mikephil.charting.formatter.ValueFormatter
+
+class ChartValueFormatter: ValueFormatter() {
+
+    override fun getAxisLabel(value: Float, axis: AxisBase?): String {
+        return super.getAxisLabel(value, axis)
+    }
+}
